+++ Assets/Scripts/Input/BaseInputUpdater.cs
@@ -1,0 +1,9 @@
+﻿using Updater;
+
+namespace Input
+{
+    public abstract class BaseInputUpdater : IUpdater
+    {
+        public abstract void Update(float deltaTime);
+    }
+}