+++ Assets/Scripts/Input/IInputView.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using UnityEngine;
+
+namespace Input
+{
+    public interface IInputView
+    {
+        event Action OnDebugPanelVisibilityChange;
+        event Action<float> OnThrust;
+        event Action<float> OnStrafe;
+        event Action<float> OnUpDown;
+        event Action<float> OnRoll;
+        event Action<Vector2> OnPitchYaw;
+        event Action OnBoost;
+        event Action OnBoostCanceled;
+        event Action<Vector2> OnMouseMove;
+        
+        bool CheckFireInputPressed();
+        bool CheckPointerOverUI();
+    }
+}