+++ Assets/Scripts/Input/Windows/IWindowsInputView.cs
@@ -1,0 +1,8 @@
+﻿namespace Input.Windows
+{
+    public interface IWindowsInputView : IInputView
+    {
+        void Initialize();
+        void Dispose();
+    }
+}