+++ Assets/Scripts/Input/Windows/WindowsInputModel.cs
@@ -1,0 +1,6 @@
+﻿namespace Input.Windows
+{
+    public class WindowsInputModel : BaseInputModel
+    {
+    }
+}